feat: Update Reminder data class
Updated the `Reminder` data class to include documentation and changed the `id` type from `Long` to `Int`.

This change improves code clarity and data representation.

+++ domain/src/main/java/city/zouitel/domain/model/Reminder.kt
@@ -2,9 +2,17 @@ package city.zouitel.domain.model
 
 import androidx.annotation.Keep
 
+/**
+ * Data class representing a reminder.
+ *
+ * @property id The unique identifier for the reminder. Defaults to 0.
+ * @property uid The unique identifier of the user who owns this reminder. Defaults to an empty string.
+ * @property atTime The time at which the reminder is scheduled to occur (in milliseconds since the epoch). Defaults to 0L.
+ * @property isPassed A boolean flag indicating whether the reminder has passed its scheduled time. Defaults to false.
+ */
 @Keep
 data class Reminder(
-    val id: Long = 0L,
+    val id: Int = 0,
     val uid: String = "",
     val atTime: Long = 0L,
     val isPassed: Boolean = false
